+++ Makefile
@@ -15,7 +15,12 @@ INC_FLAGS := $(addprefix -I,$(INC_DIRS))
 
 CPPFLAGS ?= $(INC_FLAGS) -MMD -MP -D_POSIX_C_SOURCE=200809L
 CPPFLAGS += $(PKG_CONFIG_CFLAGS)
-LDFLAGS += $(PKG_CONFIG_LIBS) -Lpdp11/build -lpdp11
+LDFLAGS += $(PKG_CONFIG_LIBS)
+
+#CPPFLAGS += -DUU_PDP11_V6
+#LDFLAGS += -Lpdp11/build -lpdp11
+CPPFLAGS += -DUU_M68K_MINIX
+LDFLAGS += -Lm68k/build -lm68k
 
 
 CFLAGS += -Wall --std=c99 -O3
