+++ src/syscall.c
@@ -19,28 +19,6 @@
 #include "../pdp11/src/util.h"
 #include "util.h"
 
-bool serializeArgvReal(machine_t *pm, int argc, char *argv[]) {
-    assert(argv[argc] == NULL);
-
-    size_t nc = 0;
-    for (int i = 0; i < argc; i++) {
-        const char *pa = argv[i];
-        do {
-            pm->args[nc++] = *pa;
-            if (nc >= sizeof(pm->args) - 1) {
-                return false;
-            }
-        } while (*pa++ != '\0');
-    }
-    if (nc & 1) {
-        pm->args[nc++] = '\0';
-    }
-
-    pm->argc = argc;
-    pm->argsbytes = nc;
-    return true;
-}
-
 static int serializeArgvVirt(machine_t *pm, uint8_t *argv) {
     uint16_t na = 0;
     uint16_t nc = 0;
