+++ src/syscall.h
@@ -2,7 +2,6 @@
 
 #include <stddef.h>
 #include <stdint.h>
-#include <stdbool.h>
 
 struct machine_tag;
 #ifndef _MACHINE_T_
@@ -10,6 +9,5 @@ struct machine_tag;
 typedef struct machine_tag machine_t;
 #endif
 
-bool serializeArgvReal(machine_t *pm, int argc, char *argv[]);
 void mysyscall(machine_t *pm);
 void syscallString(machine_t *pm, char *str, size_t size, uint8_t id);
